refactor: 검증 로직을 별도의 클래스(ChessMoveValidator)로 분리
